Implemented acidity check for iodine uptake
HOI, IONO and IONO2 should only undergo uptake to sea salt aerosol
if and when the sea salt is non-acidic, otherwise cycling on the
aerosol to produce ICl and IBr. The latter reaction was correctly
limited by alkalinity but the former reaction was not. This has now
been corrected. The sticking coefficients for HOI, IONO and IONO2
have also been corrected.

Signed-off-by: Sebastian Eastham <seastham@fas.harvard.edu>