Partitioning of HI, HBr and HCl now calculated
Halogen acids (HI, HBr and HCl) are now partitioned into fine aerosol
(IAq, BrAq and ClAq) and "other". Prior to chemistry, the quantity
of HX in gas phase, dissolved in cloud water, dissolved in a surface
layer on ice cloud particles, and dissolved into fine aerosol is
calculated based on Henry's law coefficients. This lays the groundwork
for more comprehensive aqueous-phase chemistry. As part of this update,
ISORROPIA is now called prior to gas-phase chemistry, and a new module
(aqueous_mod) is built.

Signed-off-by: Sebastian Eastham <seastham@fas.harvard.edu>