Fixed ordering of chemistry options in Makefile_header
The Makefile_header file is written in such a way that hardcoded
KPP chemistry options are required. However, these are selected by
incomplete string matching, such that "Halogensv2" is recognized
as being both "Halogens" and "Halogensv2". TO ensure that the
correct chemistry mechanism is selected, the Halogensv2 code
must follow the Halogens code, as is now the case. Otherwise, the
Halogens mechanism will always be compiled when either Halogensv2
or Halogens is specified.

Signed-off-by: Sebastian Eastham <seastham@fas.harvard.edu>