Checkpoint commit: Halogensv2 with ice recycling
This is a checkpoint update. In its current state, the code performs
minimal heterogeneous chemistry. Only HO2, NO2 and NO3 uptake, in
addition to N2O5 hydrolysis (old scheme), iodine uptake and iodine
recycling are calculated. No chemistry involving the dissolved species
scheme has yet been implemented, and warnings to this effect are displayed
if the code is run.

Signed-off-by: Sebastian Eastham <seastham@fas.harvard.edu>